Add tests for group_by variants
