+++ test/CMakeLists.txt
@@ -24,6 +24,7 @@ add_dtests(NAME test_sample_sort FILES test_sample_sort.cpp LIBS parlay)
 
 add_dtests(NAME test_primitives FILES test_primitives.cpp LIBS parlay)
 add_dtests(NAME test_random FILES test_random.cpp LIBS parlay)
+add_dtests(NAME test_group_by FILES test_group_by.cpp LIBS parlay)
 
 # -------------------------- Uninitialized memory testing ---------------------------
